Effects.System.Terminal

Description

Provides the MonadTerminal typeclass.

Since: 0.1

Effect

class Monad m => MonadTerminal m where

Represents a terminal.

Since: 0.1

Minimal complete definition

putStr, putBinary, getChar, getLine, getContents', getTerminalSize, supportsPretty

Methods

putStr :: HasCallStack => String -> m ()

Write a character to the standard output device (same as hPutChar stdout).

Since: 0.1

putStrLn :: HasCallStack => String -> m ()

The same as putStr, but adds a newline character.

Since: 0.1

putBinary :: HasCallStack => ByteString -> m ()

Write a ByteString to stdout.

Since: 0.1

getChar :: HasCallStack => m Char

Read a character from the standard input device (same as hGetChar stdin).

Since: 0.1

getLine :: HasCallStack => m String

Read a line from the standard input device (same as hGetLine stdin).

Since: 0.1

getContents' :: HasCallStack => m String

The getContents' operation returns all user input as a single string, which is fully read before being returned (same as hGetContents' stdin).

Since: 0.1

getTerminalSize :: HasCallStack => m (Window Natural)

Retrieves the terminal size.

Since: 0.1

supportsPretty :: HasCallStack => m Bool

Determines if we support ANSI styling.

Since: 0.1

Functions

print :: (HasCallStack, MonadTerminal m, Show a) => a -> m ()

The print function outputs a value of any printable type to the standard output device. Printable types are those that are instances of class Show; print converts values to strings for output using the show operation and adds a newline.

For example, a program to print the first 20 integers and their powers of 2 could be written as:

main = print ([(n, 2^n) | n <- [0..19]])

Since: 0.1

Text

putText :: (HasCallStack, MonadTerminal m) => Text -> m ()

Text version of putStr.

Since: 0.1

putTextLn :: (HasCallStack, MonadTerminal m) => Text -> m ()

Text version of putStrLn.

Since: 0.1

getTextLine :: (HasCallStack, MonadTerminal m) => m Text

Text version of getLine.

Since: 0.1

getTextContents' :: (HasCallStack, MonadTerminal m) => m Text

Text version of getContents'.

Since: 0.1

Window

getTerminalWidth :: (HasCallStack, MonadTerminal m) => m Natural

Retrieves the terminal width.

Since: 0.1

getTerminalHeight :: (HasCallStack, MonadTerminal m) => m Natural

Retrieves the terminal height.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

data Window a

Terminal window width and height

Constructors

Window

Fields

height :: !a

width :: !a

data Text

A space efficient, packed, unboxed Unicode text type.